# api/openapi.yaml
openapi: 3.1.0
info:
  title: Flex API
  version: 0.5.0
  description: API du serveur multimédia personnel Flex.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /auth/setup:
    post:
      security: []
      operationId: setupAdministrator
      summary: Crée le premier administrateur et ouvre une session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAuthRequest"
      responses:
        "201":
          description: Administrateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Flex est déjà configuré.
        "422":
          description: Informations invalides.
  /auth/login:
    post:
      security: []
      operationId: login
      summary: Ouvre une session utilisateur.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session ouverte.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Identifiants incorrects.
...
  /auth/password:
    put:
      operationId: changePassword
      summary: Modifie le mot de passe de l'utilisateur connecté et révoque ses autres sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Mot de passe modifié et session courante renouvelée.
        "401":
          description: Mot de passe actuel incorrect.
        "422":
          description: Nouveau mot de passe invalide.
  /auth/profile:
    patch:
      operationId: updateProfile
      summary: Modifie le profil de l'utilisateur connecté.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profil modifié.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Nom d’utilisateur déjà utilisé.
        "422":
          description: Nom d’utilisateur invalide.
  /users:
    get:
      operationId: listUsers
      summary: Liste les comptes Flex. Réservé aux administrateurs.
      responses:
        "200":
          description: Liste des utilisateurs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthUser"
        "403":
          description: Droits administrateur requis.
    post:
      operationId: createUser
      summary: Crée un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Utilisateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Nom d’utilisateur déjà utilisé.
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateUser
      summary: Modifie un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Utilisateur modifié.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Le dernier administrateur actif doit être conservé.
    delete:
      operationId: deleteUser
      summary: Supprime un compte et ses sessions. Réservé aux administrateurs.
      responses:
        "204":
          description: Utilisateur supprimé.
        "409":
          description: Le dernier administrateur actif doit être conservé.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flex_session
...
  schemas:
    AuthUser:
      type: object
      additionalProperties: false
      required: [id, username, role, active]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...
    SetupAuthRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          maxLength: 256
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
        role:
          type: string
          enum: [admin, user]
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [username, role, active]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...